requirements.txt
fastapi==0.110.1
pydantic==2.6.4
uvicorn==0.29.0
httpx==0.27.0
python_dotenv==1.0.1